mutasi/requirements.txt
requests==2.28.2
sqlalchemy==2.0.16
psycopg2==2.9.6
pydantic==1.10.7
fastapi==0.98.0
redis==5.0.7
pytz==2023.3
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
structlog==22.3.0
uvicorn[standard]==0.22.0